bot/requirements.txt
aiogram>=3.25.0
pymongo>=4.16.0
pydantic-settings>=2.13.0
loguru>=0.7.3
fastapi>=0.129.0
uvicorn>=0.41.0
aiohttp>=3.13.3
